backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
beautifulsoup4
httpx
